moving the functions using plugins manager from sdk => ser module
