+++ airflow/serialization/serialized_objects.py
@@ -1166,6 +1166,39 @@ class SerializedBaseOperator(BaseOperator, BaseSerialization):
         self.template_fields = BaseOperator.template_fields
         self.operator_extra_links = BaseOperator.operator_extra_links
 
+    @cached_property
+    def operator_extra_link_dict(self) -> dict[str, Any]:
+        """Returns dictionary of all extra links for the operator."""
+        op_extra_links_from_plugin: dict[str, Any] = {}
+        from airflow import plugins_manager
+
+        plugins_manager.initialize_extra_operators_links_plugins()
+        if plugins_manager.operator_extra_links is None:
+            raise AirflowException("Can't load operators")
+        for ope in plugins_manager.operator_extra_links:
+            if ope.operators and self.operator_class in ope.operators:
+                op_extra_links_from_plugin.update({ope.name: ope})
+
+        operator_extra_links_all = {link.name: link for link in self.operator_extra_links}
+        # Extra links defined in Plugins overrides operator links defined in operator
+        operator_extra_links_all.update(op_extra_links_from_plugin)
+
+        return operator_extra_links_all
+
+    @cached_property
+    def global_operator_extra_link_dict(self) -> dict[str, Any]:
+        """Returns dictionary of all global extra links."""
+        from airflow import plugins_manager
+
+        plugins_manager.initialize_extra_operators_links_plugins()
+        if plugins_manager.global_operator_extra_links is None:
+            raise AirflowException("Can't load operators")
+        return {link.name: link for link in plugins_manager.global_operator_extra_links}
+
+    @cached_property
+    def extra_links(self) -> list[str]:
+        return sorted(set(self.operator_extra_link_dict).union(self.global_operator_extra_link_dict))
+
     def get_extra_links(self, ti: TaskInstance, link_name: str) -> str | None:
         """
         For an operator, gets the URLs that the ``extra_links`` entry points to.
@@ -1326,8 +1359,7 @@ class SerializedBaseOperator(BaseOperator, BaseSerialization):
                         operator.__name__ == encoded_op["task_type"]
                         and operator.__module__ == encoded_op["_task_module"]
                     ):
-                        generic_ope = XComOperatorLink(name=ope.name, xcom_key=ope.xcom_key)
-                        op_extra_links_from_plugin.update({ope.name: generic_ope})
+                        op_extra_links_from_plugin.update({ope.name: ope})
 
             # If OperatorLinks are defined in Plugins but not in the Operator that is being Serialized
             # set the Operator links attribute
